pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "notebooklm-llm-wiki-flow"
version = "0.1.0"
description = "Portable NotebookLM → LLM Wiki → Obsidian → qmd workflow scaffold"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Moon Jeong"},
]
dependencies = [
  "typer>=0.16.0",
  "PyYAML>=6.0.2",
  "rich>=14.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
]

[project.scripts]
nlwflow = "notebooklm_llm_wiki_flow.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/notebooklm_llm_wiki_flow"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
